Fixed spotlights breaking due to float imprecision and adjusted some stuff so now it's better.

+++ engine/src/OpenGL.cpp
@@ -202,7 +202,7 @@ std::array<uint32_t, 3> OpenGL::generateGbuffer(uint32_t &gBuffer, const int32_t
     std::array<uint32_t, 3> textureIDs;
     glGenTextures(1, &textureIDs[0]);
     glBindTexture(GL_TEXTURE_2D, textureIDs[0]);
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, SCR_WIDTH, SCR_HEIGHT, 0, GL_RGBA, GL_FLOAT, NULL);
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, SCR_WIDTH, SCR_HEIGHT, 0, GL_RGBA, GL_FLOAT, NULL);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
     glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, textureIDs[0], 0);
